Converted all the old SecondIteration code to the vue components
