--- OLD_SecondIteration/README.md
@@ -1,10 +1,0 @@
-## Pre-Submit
-Pre-Submit is an application developed to help students ensure that they are handing everything in correctly for an assignment. This application is still in development!
-
-### Description
-
-
-### Requirements
-
-
-### How to Build
