--- OLD_SecondIteration/about.html
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
-        <meta charset="utf-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1">
-        <title>Pre-Submit: About</title>
-    </head>
-    <body>
-        <div class="container">
-            <div id="template"></div>
-            <div class="row h">
-                <h2>About Pre-Submit</h2>
-                <p>Pre-Submit was an idea born out of a desire to help students with requirements that are easy to miss sometimes. It is made by Avaneesh Bhat, my github page can be found <a href="">here</a>, and my <a href="https://avbhat.net">personal page</a> can be found here.</p>
-            </div>
-        </div>
-        <script src="scripts/baseTemplate.js"></script>
-        <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM" crossorigin="anonymous"></script>
-    </body>
-</html>
