--- OLD_SecondIteration/fileUpload.html
@@ -1,25 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
-        <meta charset="utf-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1">
-        <title>Pre-Submit: Upload File</title>
-    </head>
-    <body>
-        <div class="container">
-            <div id="template"></div>
-            <div class="row h">
-                <input type="file" id="fileInput" accept=".txt, .zip" multiple><br>
-                <input type="text" id="comments"><br>
-                <button onclick="uploadFile()">Check my submission</button><br>
-            </div>
-            <div class="row h">
-                <section id="results"></section>
-            </div>
-        </div>
-        <script src="scripts/baseTemplate.js"></script>
-        <script src="scripts/processResults.js"></script>
-        <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM" crossorigin="anonymous"></script>
-    </body>
-</html>
