--- OLD_SecondIteration/scripts/baseTemplate.js
@@ -1,17 +1,0 @@
-var template =`
-<div class="row h">
-    <div class="col">
-        <h1><a href="index.html">Pre-Submit</a></h1>
-        <header class="d-flex justify-content-center">
-            <ul class="nav nav-pills">
-            <li class="nav-item"><a href="index.html" class="nav-link">Index</a><br></li>
-            <li class="nav-item"><a href="templates.html" class="nav-link">Templates</a><br></li>
-            <li class="nav-item"><a href="Settings.html" class="nav-link">Settings</a><br></li>
-            <li class="nav-item"><a href="about.html" class="nav-link">About</a><br></li>
-            </ul>
-        </header>
-    </div>
-<div>
-`
-
-document.getElementById('template').innerHTML = template
