--- OLD_SecondIteration/scripts/indexTemplate.js
@@ -1,7 +1,0 @@
-var templateCode = `
-<div id="row h">            
-    <a href="templateEditor.html">Create New Template</a>
-</div>
-`
-
-document.getElementById('listTemplates').innerHTML = templateCode
