+++ src/Index.vue
@@ -1,15 +1,20 @@
 <script>
+import listTemplates from "./components/listTemplates.vue"
   export default{
     data(){
       return{
         tabIndex: 3
       }
+    },
+    components:{
+      'listTemplates': listTemplates
     }
   }
 </script>
 
 <template>
   <headerModule :activeTab="tabIndex" />
+  <listTemplates />
   <aboutBlurb v-if="tabIndex === 3" />
 
 </template>
