+++ src/components/listTemplates.vue
@@ -1,0 +1,5 @@
+<template>
+    <div id="row h">            
+        <a href="templateEditor.html">Create New Template</a>
+    </div>
+</template>
